% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimaSe.R
\name{estimaSe}
\alias{estimaSe}
\title{estimaSe}
\usage{
estimaSe(reg, y, X, params = NULL, args = NULL, weights = 1, ...)
}
\arguments{
\item{reg}{modelo estimado da classe regEM}

\item{y}{variável respostas}

\item{x}{variáveis explicativas}
}
\value{
se
}
\description{
estimaSe
}
