% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMix.R
\name{predictMix}
\alias{predictMix}
\title{predictMix}
\usage{
predictMix(
  reg,
  x = NULL,
  r = NULL,
  grupos = NULL,
  class = T,
  real = NULL,
  estimator = "mean"
)
}
\arguments{
\item{reg}{objeto de regressão de misturas}

\item{x}{variável explicativas}

\item{r}{variáveis explicativas}

\item{class}{classificar}
}
\value{
y
}
\description{
predictMix
}
