% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEM.R
\name{regEM}
\alias{regEM}
\title{regEM}
\usage{
regEM(
  y,
  x,
  g = 2,
  ...,
  tol = 1e-06,
  family = "MixNormal",
  grupoReal = NULL,
  max_iter = 1000,
  min_iter = 1,
  verbose = F,
  showSE = F
)
}
\arguments{
\item{y}{variável respostas}

\item{x}{variáveis explicativas}

\item{g}{número de grupos}

\item{tol}{tolerância}

\item{family}{família}

\item{grupoReal}{grupo real da classificação}

\item{max_iter}{número máximo de iterações}

\item{min_iter}{número mínimo de iterações}

\item{verbose}{mostrar a verossimilhança passo a passo}
}
\value{
resultados finais
}
\description{
regEM
}
