% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMix.R
\name{predictMix}
\alias{predictMix}
\title{predictMix}
\usage{
predictMix(reg, x = NULL, r = NULL, class = T)
}
\arguments{
\item{reg}{objeto de regressão de misturas}

\item{x}{variável explicativas}

\item{r}{variáveis explicativas}

\item{class}{classificar}
}
\value{
y
}
\description{
predictMix
}
